% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_functions.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with the IUROPA API}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{A string. Your IUROPA username.}

\item{password}{A string. Your IUROPA password.}
}
\value{
This function returns an object of class \code{iuropa_session}.
}
\description{
This function allows you to authenticate with the IUROPA API. It
  creates a session. You need to provide your IUROPA username and password to
  authenticate. This function returns an object of class
  \code{iuropa_session} that contains the credentials necessary to request
  password-protected data from the IUROPA API. You need to pass this session
  object into any function that requires a \code{session} argument. A session
  times out after 30 minutes. If your session times out, you will need to run
  this function again to create a new session. You can have multiple active
  sessions at the same time.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)}

}
