% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_functions.R
\name{check_authentication}
\alias{check_authentication}
\title{Check authentication with the IUROPA API}
\usage{
check_authentication(session)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}.}
}
\value{
This function prints a message to the console but does not return an
  object.
}
\description{
This function checks whether you are currently authenticated
  with the IUROPA API. You need to provide an object of class
  \code{iuropa_session} created by the function \code{authenticate()}.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

check_authentication(session)}

}
