% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_functions.R
\name{describe_datasets}
\alias{describe_datasets}
\title{Describe datasets in a IUROPA component}
\usage{
describe_datasets(session, component)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}.}

\item{component}{A string. The name of a IUROPA component. Use
\code{list_components()} to get a list of valid values.}
}
\value{
This function returns a tibble containing 5 variables. There is one
  observation per dataset in the specified component.
}
\description{
This function provides descriptions of all of the datasets that
  are currently available in a particular IUROPA component. You have to
  specify a IUROPA component.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- describe_datasets(
  session = session,
  component = "cjeu_database_platform"
)}

}
