% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_functions.R
\name{describe_tables}
\alias{describe_tables}
\title{Describe the tables in a component}
\usage{
describe_tables(component, session = NULL)
}
\arguments{
\item{component}{A string. The code for a component. Use
\code{list_components()} to get a list of valid values.}

\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}. This argument is only required for content that is
not yet publicly available.}
}
\value{
This function returns a tibble. There is one observation per table in
  the specified component.
}
\description{
This function provides descriptions of the tables that are
  currently available in a particular component of the IUROPA CJEU Database.
  You have to specify the code for a component. The values returned by
  \code{list_components()} are the values that are valid for this argument.
}
\examples{
\dontrun{
data <- describe_tables(
  component = "cjeu_database_platform"
)}

}
