% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets in a IUROPA component}
\usage{
list_datasets(session, component)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}.}

\item{component}{A string. The name of a IUROPA component. Use
\code{list_components()} to get a list of valid values.}
}
\value{
This function returns a string vector containing the names of the
  datasets in the specified component.
}
\description{
This function lists all of the datasets that are currently
  available in a particular IUROPA component. You have to specify a IUROPA
  component. A number of functions in the \code{iuropa} package have a
  \code{dataset} argument. The values returned by \code{list_datasets()} are
  those that are valid for this argument.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- list_datasets(
  session = session,
  component = "cjeu_database_platform"
)}

}
