% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_tables}
\alias{list_tables}
\title{List the tables in a component}
\usage{
list_tables(session = NULL, component)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}. This argument is only required for components that
are not yet publicly available.}

\item{component}{A string. The code for a component. Use
\code{list_components()} to get a list of valid values.}
}
\value{
This function returns a string vector containing the names of the
  tables in the specified component.
}
\description{
This function lists the tables that are currently available in a
  particular component of the IUROPA CJEU Database. You have to specify the
  code for a component. The values returned by \code{list_components()} are
  the values that are valid for this argument. A number of functions in the
  \code{iuropa} package have a \code{tables} argument. The values returned by
  \code{list_tables()} are the values that are valid for this argument.
}
\examples{
\dontrun{
out <- list_tables(
  component = "cjeu_database_platform"
)}

}
