% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_variables}
\alias{list_variables}
\title{List variables in a IUROPA dataset}
\usage{
list_variables(session, component, dataset)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}.}

\item{component}{A string. The name of a IUROPA component. Use
\code{list_components()} to get a list of valid values.}

\item{dataset}{A string. The name of a IUROPA dataset in the specified IUROPA
component. Use \code{list_datasets()} to get a list of valid values.}
}
\value{
This function returns a string vector containing the names of the
  variables in the specified dataset.
}
\description{
This function lists all of the variables that are currently
  available in a particular IUROPA dataset. You have to specify a IUROPA
  component and a IUROPA dataset in that component.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- list_variables(
  session = session,
  component = "cjeu_database_platform",
  dataset = "decisions"
)}

}
