% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_citation.R
\name{print_citation}
\alias{print_citation}
\title{Print the citation for a IUROPA component}
\usage{
print_citation(component, version)
}
\arguments{
\item{component}{A string. The name of a IUROPA component. Use
\code{list_components()} to get a list of valid values.}

\item{version}{A string. A version number in the format \code{#.#}, such as
\code{0.1}.}
}
\description{
This function generates the recommended citations for a IUROPA component. You
need to provide a component id and a version number. The version number
should be the version number for the IUROPA component, not the version number
of the \code{R} package.
}
\examples{
\dontrun{
print_citation(
  component = "cjeu_database_platform",
  version = "0.1"
)}

}
