% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_functions.R
\name{describe_components}
\alias{describe_components}
\title{Describe the components in the IUROPA CJEU Database}
\usage{
describe_components(session = NULL)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}. This argument is only required for content that is
not yet publicly available.}
}
\value{
This function returns a tibble. There is one observation per
  component.
}
\description{
The IUROPA CJEU Database includes a number of components. This
  function provides descriptions of the components that are currently
  available via the IUROPA API.
}
\examples{
\dontrun{
data <- describe_components()}

}
