% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_functions.R
\name{describe_variables}
\alias{describe_variables}
\title{Describe variables in a IUROPA dataset}
\usage{
describe_variables(session, component, dataset)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}.}

\item{component}{A string. The name of a IUROPA component. Use
\code{list_components()} to get a list of valid values.}

\item{dataset}{A string. The name of a IUROPA dataset in the specified IUROPA
component. Use \code{list_datasets()} to get a list of valid values.}
}
\value{
This function returns a tibble containing 7 variables. There is one
  observation per variable in the specified dataset.
}
\description{
This function provides descriptions of all of the variables that
  are currently available in a particular IUROPA dataset. You have to specify
  a IUROPA component and a IUROPA dataset in that component.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- describe_variables(
  session = session,
  component = "cjeu_database_platform",
  dataset = "decisions"
)}

}
