% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_components}
\alias{list_components}
\title{List the components in the IUROPA CJEU Database}
\usage{
list_components(session = NULL)
}
\arguments{
\item{session}{An object of class \code{iuropa_session} created by
\code{authenticate()}. This argument is only required for content that is
not yet publicly available.}
}
\value{
This function returns a string vector containing the names of the
  components that are currently available via the IUROPA API.
}
\description{
The IUROPA CJEU Database includes a number of components. This
  function lists the components that are currently available via the IUROPA
  API. A number of functions in the \code{iuropa} package have a
  \code{component} argument. The values returned by \code{list_components()}
  are the values that are valid for this argument.
}
\examples{
\dontrun{
data <- list_components()}

}
