% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_citation.R
\name{print_citation}
\alias{print_citation}
\title{Print the citation for a IUROPA component}
\usage{
print_citation(component)
}
\arguments{
\item{component}{A string. The code for a component. Use
\code{list_components()} to get a list of valid values.}
}
\description{
This function generates the recommended citations for a IUROPA component. You
need to provide the code for a component.
}
\examples{
\dontrun{
print_citation(
  component = "cjeu_database_platform"
)}

}
